refactor: use the cli-server-api package (#1140)
* refactor: use the cli-server-api package

* Rename `server` folder to `start`

* Fix debugger always opening
