+++ packages/cli/package.json
@@ -30,17 +30,15 @@
   "dependencies": {
     "@hapi/joi": "^15.0.3",
     "@react-native-community/cli-debugger-ui": "^4.8.0",
+    "@react-native-community/cli-server-api": "^4.8.0",
     "@react-native-community/cli-tools": "^4.8.0",
     "@react-native-community/cli-types": "^4.8.0",
     "chalk": "^3.0.0",
     "command-exists": "^1.2.8",
     "commander": "^2.19.0",
-    "compression": "^1.7.1",
-    "connect": "^3.6.5",
     "cosmiconfig": "^5.1.0",
     "deepmerge": "^3.2.0",
     "envinfo": "^7.1.0",
-    "errorhandler": "^1.5.0",
     "execa": "^1.0.0",
     "find-up": "^4.1.0",
     "fs-extra": "^8.1.0",
@@ -63,17 +61,14 @@
     "serve-static": "^1.13.1",
     "strip-ansi": "^5.2.0",
     "sudo-prompt": "^9.0.0",
-    "wcwidth": "^1.0.1",
-    "ws": "^1.1.0"
+    "wcwidth": "^1.0.1"
   },
   "peerDependencies": {
     "react-native": "^0.62.0-rc.0"
   },
   "devDependencies": {
     "@types/command-exists": "^1.2.0",
-    "@types/compression": "^1.0.1",
     "@types/cosmiconfig": "^5.0.3",
-    "@types/errorhandler": "^0.0.32",
     "@types/fs-extra": "^8.1.0",
     "@types/glob": "^7.1.1",
     "@types/graceful-fs": "^4.1.3",
@@ -83,7 +78,6 @@
     "@types/mkdirp": "^0.5.2",
     "@types/semver": "^6.0.2",
     "@types/wcwidth": "^1.0.0",
-    "@types/ws": "^6.0.3",
     "slash": "^3.0.0",
     "snapshot-diff": "^0.7.0"
   }
