--- packages/cli/src/commands/server/middleware/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>React Native</title>
-  </head>
-  <body>
-    <p>React Native packager is running.</p>
-    <p><a href="https://reactnative.dev">Visit documentation</a></p>
-  </body>
-</html>
