--- packages/cli/src/commands/server/middleware/indexPage.ts
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-import http from 'http';
-import fs from 'fs';
-import path from 'path';
-
-export default function indexPageMiddleware(
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-  next: (err?: any) => void,
-) {
-  if (req.url === '/') {
-    res.setHeader('Content-Type', 'text/html');
-    res.end(fs.readFileSync(path.join(__dirname, 'index.html')));
-  } else {
-    next();
-  }
-}
