--- packages/cli/src/commands/server/middleware/statusPageMiddleware.ts
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-import http from 'http';
-
-/**
- * Status page so that anyone who needs to can verify that the packager is
- * running on 8081 and not another program / service.
- */
-export default function statusPageMiddleware(
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-  next: (err?: any) => void,
-) {
-  if (req.url === '/status') {
-    res.end('packager-status:running');
-  } else {
-    next();
-  }
-}
