+++ packages/cli/tsconfig.json
@@ -7,6 +7,7 @@
   "references": [
     {"path": "../tools"},
     {"path": "../cli-types"},
-    {"path": "../debugger-ui"}
+    {"path": "../debugger-ui"},
+    {"path": "../cli-server-api"}
   ]
 }
